+++ vrp-pragmatic/src/format/solution/writer.rs
@@ -301,6 +301,7 @@ fn map_code_reason(code: i32) -> (i32, &'static str) {
         PRIORITY_CONSTRAINT_CODE => (105, "cannot be served due to priority"),
         AREA_CONSTRAINT_CODE => (106, "cannot be assigned due to area constraint"),
         DEPOT_CONSTRAINT_CODE => (107, "cannot be assigned due to vehicle dispatch"),
+        TOUR_SIZE_CONSTRAINT_CODE => (108, "cannot be assigned due to tour size constraint of vehicle"),
         _ => (0, "unknown"),
     }
 }