+++ vrp-pragmatic/tests/unit/format/problem/reader_test.rs
@@ -174,7 +174,12 @@ fn can_read_complex_problem() {
                 }],
                 capacity: vec![10, 1],
                 skills: Some(vec!["unique1".to_string(), "unique2".to_string()]),
-                limits: Some(VehicleLimits { max_distance: Some(123.1), shift_time: Some(100.), allowed_areas: None }),
+                limits: Some(VehicleLimits {
+                    max_distance: Some(123.1),
+                    shift_time: Some(100.),
+                    tour_size: Some(3),
+                    allowed_areas: None,
+                }),
             }],
             profiles: create_default_profiles(),
         },